+++ Sprint 2/app/templates/auth/selection_type.html
@@ -1,8 +1,8 @@
 {% extends "base.html" %} {% block title %}Jobs{% endblock %} 
 {% block content%}
       <div class="group">
-        <div class="item" style="background-image: url(https://images.unsplash.com/photo-1544654262-e295983be0f4?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxzZWFyY2h8OXx8cG9ydHJhaXR8ZW58MHx8MHx8&auto=format&fit=crop&w=800&q=60)"><a href='/sign-up'><h3>Seeker</h3></a></div>
-        <div class="item" style="background-image: url(https://images.unsplash.com/photo-1670272505284-8faba1c31f7d?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxzZWFyY2h8MjB8fHBvcnRyYWl0fGVufDB8fDB8fA%3D%3D&auto=format&fit=crop&w=800&q=60)"><a href='/signup_employer'><h3>Employer</h3></a></div>
-        <div class="item" style="background-image: url(https://images.unsplash.com/photo-1632406895715-c9521b447fab?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxzZWFyY2h8OTN8fHBvcnRyYWl0fGVufDB8fDB8fA%3D%3D&auto=format&fit=crop&w=800&q=60)"><a href='/signup_admin'><h3>Admin</h3></a></div>
+        <div class="item" style="background-image: url(https://images.unsplash.com/photo-1544654262-e295983be0f4?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxzZWFyY2h8OXx8cG9ydHJhaXR8ZW58MHx8MHx8&auto=format&fit=crop&w=800&q=60)"><a href='/auth/sign-up'><h3>Seeker</h3></a></div>
+        <div class="item" style="background-image: url(https://images.unsplash.com/photo-1670272505284-8faba1c31f7d?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxzZWFyY2h8MjB8fHBvcnRyYWl0fGVufDB8fDB8fA%3D%3D&auto=format&fit=crop&w=800&q=60)"><a href='/auth/signup_employer'><h3>Employer</h3></a></div>
+        <div class="item" style="background-image: url(https://images.unsplash.com/photo-1632406895715-c9521b447fab?ixlib=rb-4.0.3&ixid=MnwxMjA3fDB8MHxzZWFyY2h8OTN8fHBvcnRyYWl0fGVufDB8fDB8fA%3D%3D&auto=format&fit=crop&w=800&q=60)"><a href='/auth/signup_admin'><h3>Admin</h3></a></div>
       </div>
     {% endblock %}
