+++ Sprint 2/app/templates/auth/signup.html
@@ -3,9 +3,9 @@
 
 {% block content %}
 <h1>Job Seeker Sign Up Page</h1>
-<a id="signup_employer" href="/signup_employer">I am an employer.</a>
+<a id="signup_employer" href="/auth/signup_employer">I am an employer.</a>
 <br>
-<a id="sign-up" href="/signup_admin">I am an admin.</a>
+<a id="sign-up" href="/auth/signup_admin">I am an admin.</a>
 <form method="POST">
     <h3 align="center">Sign Up</h3>
     <div class="form-group">
