+++ Sprint 2/app/templates/auth/signup_admin.html
@@ -3,9 +3,9 @@
 
 {% block content %}
 <h1>Admin Sign Up Page</h1>
-<a id="sign-up" href="/sign-up">I am a job seeker.</a>
+<a id="sign-up" href="/auth/sign-up">I am a job seeker.</a>
 <br>
-<a id="signup_employer" href="/sign-up">I am an employer.</a>
+<a id="signup_employer" href="/auth/sign-up">I am an employer.</a>
 <form method="POST">
     <h3 align="center">Sign Up</h3>
     <div class="form-group">
