+++ Sprint 2/app/templates/auth/signup_employer.html
@@ -3,9 +3,9 @@
 
 {% block content %}
 <h1>Employer Sign Up Page</h1>
-<a id="sign-up" href="/sign-up">I am a job seeker.</a>
+<a id="sign-up" href="/auth/sign-up">I am a job seeker.</a>
 <br>
-<a id="sign-up" href="/signup_admin">I am an admin.</a>
+<a id="sign-up" href="/auth/signup_admin">I am an admin.</a>
 <form method="POST">
     <h3 align="center">Sign Up</h3>
     <div class="form-group">
