+++ Sprint 2/app/templates/contact.html
@@ -26,7 +26,7 @@
 
     <body>
         <div class="container">
-            <form action="{{ url_for('auth.contact') }}" method=post>
+            <form action="{{ url_for('main.contact') }}" method=post>
                 <h1>Need help? Contact us here. </h1>
                 {{ form.hidden_tag() }}
                 {{ form.name(class_="input", placeholder="Name") }}
